requirements.txt
langchain
langchain_pinecone
langchain_community
pinecone
sentence-transformers
pypdf
faiss-cpu
ollama
pandas
nltk
rouge-score
langchain_openai
matplotlib
sacrebleu
evaluate